Libft, printf, gnl

+++ libft/ft_substr.c
@@ -6,17 +6,17 @@
 /*   By: levincen <levincen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 13:25:46 by levincen          #+#    #+#             */
-/*   Updated: 2024/11/11 10:51:59 by levincen         ###   ########.fr       */
+/*   Updated: 2024/12/30 13:25:11 by levincen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, int start, int len)
 {
 	char	*sub;
-	size_t	i;
+	int	i;
 
 	if (!s)
 		return (NULL);
